Two new base directories are created: /src and /test. /src represents everything that goes into the app, while /test runs unit tests on those files.
A new class Card is created. One can create a new card with the *new* method, given a suit and a value. TODO: restrict which suits and values are possible using the suits and values tables. There is another method, *beats*, which when called on a card, returns whether that card beats another one, given a vira.

A new file test.lua is created. It allows to define unit tests. Unit tests are to be loaded from the test.cfg file. To run all test, simply call run this file on the command line. For every failed test, a message will appear containing which test failed, which file is responsible for that test and why the test failed. At the end, a message will show how many test have passed out of the total.

To define a unit test, simply create a test function, followed by a table. The table needs three attributes. The attribute *description*, which gives a brief explanation of what it is going to be tested, *expected*, the expected result of the test, and *result*, the function that actually runs the test, returing the result.